Extract kafkaConfig to root dir
