+++ consumer/consumer.js
@@ -1,6 +1,6 @@
 const writeToDB = require('./writeToDB');
 const { Consumer } = require('sinek');
-const kafkaConfig = require('./kafkaConfig');
+const kafkaConfig = require('../kafkaConfig');
 const consumer = new Consumer('events', kafkaConfig);
 const withBackpressure = true;
